+++ src/test/java/io/github/astrapi69/icon/IconExtensionsTest.java
@@ -36,6 +36,7 @@ import javax.swing.UIManager;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
 import io.github.astrapi69.file.delete.DeleteFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
@@ -67,7 +68,7 @@ class IconExtensionsTest
 	@Disabled("because this test case fails on github actions. Locally it runs successfully.")
 	void storeIcon() throws IOException
 	{
-		final File imgDir = new File(PathFinder.getSrcTestResourcesDir(), "img");
+		File imgDir = new File(PathFinder.getSrcTestResourcesDir(), "img");
 
 		File xmas = new File(imgDir, "xmas");
 		File outputFile = new File(xmas, "fileView.png");
@@ -77,6 +78,16 @@ class IconExtensionsTest
 		assertTrue(outputFile.exists());
 		DeleteFileExtensions.delete(outputFile);
 
+	}
+
+	/**
+	 * Test method for {@link IconExtensions} with {@link BeanTester}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(IconExtensions.class);
 	}
 
 }